Do not re-use return arguments when waiting for select block return
In some parts of the code we make use of a select block to wait for the result of the target operation, or return a "context deadline exceeded" error if the provided context expires.

By reusing the output parameters, we were causing nasty data races, as even if the context got cancelled, the operation can be fulfilled after that, reusing the output parameters.

 This commit fixes that, by using a response struct and a response channel.